Split SpatialCanvas barrel and widen shape state types
