+++ packages/layers/src/index.ts
@@ -14,7 +14,6 @@ export {
   EMPTY_SHAPE_FEATURE_STATE_RUNTIME,
   isShapeFeatureStateRuntime,
   normalizeShapeFeatureState,
-  type ShapeFeatureStateRuntime,
   resolveShapeFeatureFromPick,
   resolveShapeFeatureFromPickInfo,
   resolveShapeTooltipFromPickInfo,
@@ -22,7 +21,10 @@ export {
   type ShapesLayerPickEvent,
   type ShapeCircleRenderDatum,
   type ShapeFeatureRenderDatum,
+  type ShapeFeatureStateInput,
+  type ShapeFeatureStateRuntime,
   type ShapePolygonRenderDatum,
+  type SpatialShapesRuntimeSublayer,
   type ShapesPrebuiltData,
   type ShapesRenderDataLike,
   type ShapeStrokeWidthUnits,
