+++ packages/vis/src/SpatialCanvas/public.ts
@@ -1,0 +1,32 @@
+export {
+  SpatialFeatureTooltip,
+  type SpatialCanvasTooltipRenderProps,
+  type SpatialFeatureTooltipData,
+  type SpatialFeatureTooltipItem,
+  type SpatialFeatureTooltipProps,
+  type SpatialFeatureTooltipSection,
+} from './SpatialFeatureTooltip';
+export {
+  SpatialCanvasProvider,
+  useSpatialCanvasActions,
+  useSpatialCanvasStore,
+  useSpatialCanvasStoreApi,
+} from './context';
+export { useSpatialViewState, useViewStateUrl } from './hooks';
+export { createSpatialCanvasStore } from './stores';
+export type { SpatialCanvasStoreApi } from './stores';
+export type * from './types';
+export { VivSpatialViewer } from './VivSpatialViewer';
+export {
+  composeSpatialDeckLayers,
+  shouldAutoFitSpatialView,
+  shouldRenderInternalTooltip,
+  SpatialCanvasViewer,
+  useSpatialCanvasRenderer,
+} from './SpatialCanvasViewer';
+export type {
+  SpatialCanvasViewerProps,
+  SpatialCanvasViewerRenderTooltip,
+} from './SpatialCanvasViewer';
+export type { SpatialCanvasProps } from './index';
+export type { ImageLayerConfig as VivImageLayerConfig } from './useLayerData';
