+++ packages/vis/src/index.ts
@@ -27,7 +27,7 @@ export {
   shouldRenderInternalTooltip,
   shouldAutoFitSpatialView,
   useSpatialCanvasRenderer,
-} from './SpatialCanvas';
+} from './SpatialCanvas/public';
 export type {
   SpatialCanvasStoreApi,
   SpatialCanvasState,
@@ -46,5 +46,5 @@ export type {
   SpatialFeatureTooltipSection,
   SpatialCanvasTooltipRenderProps,
   SpatialFeatureTooltipProps,
-} from './SpatialCanvas';
-export { SpatialFeatureTooltip } from './SpatialCanvas';
+} from './SpatialCanvas/public';
+export { SpatialFeatureTooltip } from './SpatialCanvas/public';
